Better customizable emitting material
Instead of applying a tint to the emission parts of a texture directly
in code, use a shader node for this task so it can be tweaked much
simpler in Blender node editor. The fullbright texture only describes
the percentage of emission now, giving a better seperation of concerns.
That also allows a simpler node setup using a principled BSDF shader
with both diffuse and emissive component.

Also allows models to have tinted emission.