Reduce noise of fullbright overlay objects
It seems like conditionally changing emission strength based on if it
is a camera ray or not messes with the emission sampling, causing quite
a lot of noise. By making the overlay objects entirely the fixed net
emission strength and hiding them from the camera and using the camera
emission strength on the underlying "underlay" material we can avoid
that.